# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample API starter by fiber.",
        "title": "Golang Fiber API starter",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:7000",
    "basePath": "/api",
    "paths": {
...
        "/logs": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "get Logs",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "logs"
                ],
                "summary": "List Logs",
                "parameters": [
                    {
                        "type": "number",
                        "example": 2,
                        "description": "id",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "example": 2,
                        "description": "search by userId",
                        "name": "userId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "go practice",
                        "description": "search by task",
                        "name": "task",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "example": true,
                        "description": "search by done",
                        "name": "done",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "1",
                        "description": "page number for pagination",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "10",
                        "description": "items per page for pagination",
                        "name": "items",
                        "in": "query"
                    }
                ],
                "responses": {}
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create log(s)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "logs"
                ],
                "summary": "Create new log(s)",
                "parameters": [
                    {
                        "description": "single Log request json",
                        "name": "Log",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "{ \"task\": \"take shower\", \"done\": false }"
                        }
                    },
                    {
                        "description": "batch Log request json",
                        "name": "Logs",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "[{ \"task\": \"take shower\", \"done\": false }, { \"task\": \"go practice\", \"done\": false }]"
                        }
                    }
                ],
                "responses": {}
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "delete log(s)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "logs"
                ],
                "summary": "Delete log(s)",
                "parameters": [
                    {
                        "description": "array of log IDs",
                        "name": "logIds",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "[1,2,3]"
                        }
                    }
                ],
                "responses": {}
            },
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update log(s)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "logs"
                ],
                "summary": "Update existing log(s)",
                "parameters": [
                    {
                        "description": "single Log request json",
                        "name": "Log",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "{ \"id\":2, \"task\": \"take shower\", \"done\": false }"
                        }
                    },
                    {
                        "description": "batch Log request json",
                        "name": "Logs",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "[{ \"id\":2, \"task\": \"take shower\", \"done\": false, createdAt: \"2021-01-11\" }, { \"id\":3, \"task\": \"go practice\", \"done\": false }]"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/logs/{logId}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "get Log by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "logs"
                ],
                "summary": "Get Log by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "example": 12,
                        "description": "log ID",
                        "name": "logId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}